internal/e2e: Move e2e tests into featuretests.
Closes #2549.
